requirements.txt
streamlit
pillow
opencv-python
numpy
torch
docarray
facenet-pytorch